services/etl/requirements.txt
clickhouse-driver==0.2.4
pydantic==1.9.1
orjson==3.8.0
kafka-python==2.0.2
backoff==2.1.2
